test: split associated field tests (#2597)
